requirements.dev.txt
# dev dependencies
ipython>=7.0.0, <=8.32.0
ipdb
pudb
pytest-watch

# Install project as editable
-e .
